Refactor schemas
Let's split the schemas used by the HTTP API and the database. That way,
we can have a domain model that's separate from either of these.

At the moment, this is arguably just verbose and noisy; however, it
should reveal its benefits as Tabbit becomes more complex. This decision
can be revisited in the future.