Refactor schemas (#18)
Let's split the schemas used by the HTTP API and the database. That way,
we can have a domain model that's separate from either of these.

At the moment, this is arguably just verbose and noisy; however, it
should reveal its benefits as Tabbit becomes more complex. This decision
can be revisited in the future.

I did try migrating from Pydantic to built-in dataclasses for the
database schemas, but there isn't a nice way to port over the unset vs
`None` distinction that Pydantic just handles for us. Might revisit.